requirements.txt
fastapi>=0.100.0
sqlalchemy>=1.4.0
pydantic-settings>=2.0.0
passlib>=1.7.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.0